building contents for v26
This downgrades pyOpenSSL to 0.12 (zooko says it has compatibility
problsm with libssl-0.9.8).

setuptools_trial and setuptools_darcs are still here. Tahoe doesn't
use them anymore, but some of the embedded dependencies (zfec,
zbase32, pyutil) might, and I'm more interested in getting this
updated quickly than saving the space of those two packages.

This copy uses the upstream tarballs for everything, some of which
are .bz2 or .zip